Update site style, include noscript and fix formatting
The layout has been (somewhat) redesigned to flow better between mobile
and desktop. Also, the navigation bar contains a search bar too which
will be used in the future for easy context-based searching.
As it is right now, user accounts still need to be implemented, and
pretty much everything else which actually has some function.