More work on logging in and user accounts...
User accounts are still incomplete at the moment, however they are
becoming more complete. We still need to finish:

* Registration and account creation
* Verification (i.e. checking they are who they say they are)
* Roles (Admin, Moderator, User, etc.)
* And all the other goodies with come along with user accounts

Also, I was talking to comp09 in the gitter chat and he said that we
should use PHPass. I took a look at the project page and while it seems
cool, I'm not sure that we should use it right now, mostly because it will
take slightly longer to implement. I would like some more opinions on
this.

-DigiTechs